write install info
